app/requirement.txt
fastapi==0.119.1
uvicorn==0.38.0
torch==2.8.0
transformers==4.57.1
faiss-cpu==1.8.0
motor==3.7.1
pydantic==2.10.1
pydantic-settings==2.11.0
numpy==1.26.4
Pillow==11.3.0
requests==2.32.5
